Fix some CI build issues

+++ examples/all-clusters-app/all-clusters-common/src/WhmDelegateImpl.cpp
@@ -316,11 +316,11 @@ uint16_t WaterHeaterManagementDelegate::GetActiveTargetWaterTemperature() const
 
 uint8_t WaterHeaterManagementDelegate::CalculateTankPercentage() const
 {
-    int16_t tankPercentage = 100 * (static_cast<int16_t>(mWaterTemperature) - static_cast<int16_t>(mColdWaterTemperature)) /
-        (static_cast<int16_t>(GetActiveTargetWaterTemperature()) - static_cast<int16_t>(mColdWaterTemperature));
+    int32_t tankPercentage = 100 * (static_cast<int32_t>(mWaterTemperature) - static_cast<int32_t>(mColdWaterTemperature)) /
+        (static_cast<int32_t>(GetActiveTargetWaterTemperature()) - static_cast<int32_t>(mColdWaterTemperature));
 
-    tankPercentage = std::min(tankPercentage, static_cast<int16_t>(100));
-    tankPercentage = std::max(tankPercentage, static_cast<int16_t>(0));
+    tankPercentage = std::min(tankPercentage, static_cast<int32_t>(100));
+    tankPercentage = std::max(tankPercentage, static_cast<int32_t>(0));
 
     return static_cast<uint8_t>(tankPercentage);
 }
@@ -356,7 +356,7 @@ void WaterHeaterManagementDelegate::DrawOffHotWater(Percent percentageReplaced, 
 {
     // First calculate the new average water temperature
     mWaterTemperature =
-        (mWaterTemperature * static_cast<uint16_t>(100 - percentageReplaced) + replacedWaterTemperature * percentageReplaced) / 100;
+        static_cast<uint16_t>(mWaterTemperature * (100 - percentageReplaced) + replacedWaterTemperature * percentageReplaced) / 100;
 
     // Replaces percentageReplaced% of the water in the tank with water of a temperature replacedWaterTemperature
     // Only supported if the kTankPercent feature is supported.
